requirements.txt
filetype
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
selenium